+++ ios/flutter_defender/Sources/flutter_defender/DefenderMessages.g.swift
@@ -563,6 +563,7 @@ protocol DefenderFlutterApiProtocol {
   func onScreenCaptureChanged(active activeArg: Bool, completion: @escaping (Result<Void, PigeonError>) -> Void)
   func onOverlayViolation(completion: @escaping (Result<Void, PigeonError>) -> Void)
   func onForegroundStateChanged(active activeArg: Bool, completion: @escaping (Result<Void, PigeonError>) -> Void)
+  func onWindowFocusChanged(hasFocus hasFocusArg: Bool, completion: @escaping (Result<Void, PigeonError>) -> Void)
 }
 class DefenderFlutterApi: DefenderFlutterApiProtocol {
   private let binaryMessenger: FlutterBinaryMessenger
@@ -632,6 +633,24 @@ class DefenderFlutterApi: DefenderFlutterApiProtocol {
     let channelName: String = "dev.flutter.pigeon.flutter_defender.DefenderFlutterApi.onForegroundStateChanged\(messageChannelSuffix)"
     let channel = FlutterBasicMessageChannel(name: channelName, binaryMessenger: binaryMessenger, codec: codec)
     channel.sendMessage([activeArg] as [Any?]) { response in
+      guard let listResponse = response as? [Any?] else {
+        completion(.failure(createConnectionError(withChannelName: channelName)))
+        return
+      }
+      if listResponse.count > 1 {
+        let code: String = listResponse[0] as! String
+        let message: String? = nilOrValue(listResponse[1])
+        let details: String? = nilOrValue(listResponse[2])
+        completion(.failure(PigeonError(code: code, message: message, details: details)))
+      } else {
+        completion(.success(()))
+      }
+    }
+  }
+  func onWindowFocusChanged(hasFocus hasFocusArg: Bool, completion: @escaping (Result<Void, PigeonError>) -> Void) {
+    let channelName: String = "dev.flutter.pigeon.flutter_defender.DefenderFlutterApi.onWindowFocusChanged\(messageChannelSuffix)"
+    let channel = FlutterBasicMessageChannel(name: channelName, binaryMessenger: binaryMessenger, codec: codec)
+    channel.sendMessage([hasFocusArg] as [Any?]) { response in
       guard let listResponse = response as? [Any?] else {
         completion(.failure(createConnectionError(withChannelName: channelName)))
         return